server/requirements.txt
alembic==1.17.2
fastapi==0.124.2
python-dotenv==1.2.1
SQLAlchemy==2.0.45
sqlmodel==0.0.27
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
asyncpg==0.29.0
